requirements.txt
annotated-types==0.7.0
anyio==4.10.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
ecdsa==0.19.1
fastapi==0.116.1
google-auth==2.40.3
google-auth-oauthlib==1.2.2
gspread==6.2.1
h11==0.16.0
httptools==0.6.4
idna==3.10
MarkupSafe==3.0.2
nexus-rpc==1.1.0
oauthlib==3.3.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
python-dotenv==1.1.1
python-http-client==3.3.7
PyYAML==6.0.2
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
sendgrid==6.12.4
six==1.17.0
sniffio==1.3.1
starlette==0.47.3
temporalio==1.17.0
types-protobuf==6.30.2.20250822
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.3
